+++ include/common/monolish_common.hpp
@@ -168,6 +168,54 @@ template <typename T> matrix::COO<T> frank_matrix(const int &M);
  **/
 template <typename T> T frank_matrix_eigenvalue(const int &M, const int &N);
 
+/**
+ * @brief create tridiagonal Toeplitz matrix
+ * @param M # of row and col
+ * @param a value of diagonal elements
+ * @param b value of next-to-diagonal elements
+ * @note
+ * - # of computation: M
+ * - Multi-threading: false
+ * - GPU acceleration: false
+ **/
+template <typename T>
+matrix::COO<T> tridiagonal_toeplitz_matrix(const int &M, T a, T b);
+
+/**
+ * @brief Nth smallest eigenvalue of MxM tridiagonal Toeplitz matrix
+ * @param M dimension of tridiagonal Toeplitz matrix
+ * @param N #-th eigenvalue from the bottom
+ * @param a value of diagonal elements
+ * @param b value of next-to-diagonal elements
+ * @note
+ * - # of computation: O(1)
+ * - Multi-threading: false
+ * - GPU acceleration: false
+ **/
+template <typename T>
+T tridiagonal_toeplitz_matrix_eigenvalue(const int &M, int N, T a, T b);
+
+/**
+ * @brief create 1D Laplacian matrix
+ * @param M # of row and col
+ * @note
+ * - # of computation: M
+ * - Multi-threading: false
+ * - GPU acceleration: false
+ **/
+template <typename T> matrix::COO<T> laplacian_matrix_1D(const int &M);
+
+/**
+ * @brief Nth smallest eigenvalue of 1D Laplacian matrix
+ * @param M dimension of tridiagonal Toeplitz matrix
+ * @param N #-th eigenvalue from the bottom
+ * @note
+ * - # of computation: O(1)
+ * - Multi-threading: false
+ * - GPU acceleration: false
+ **/
+template <typename T> T laplacian_matrix_1D_eigenvalue(const int &M, int N);
+
 // send///////////////////
 
 /**