+++ src/blas/matrix/copy/d2f.sh
@@ -1,0 +1,3 @@
+cat double_copy.cpp \
+    | sed -e 's/double/float/g' \
+    > float_copy.cpp