+++ src/utils/coo_util.cpp
@@ -92,8 +92,33 @@ template <typename T> COO<T>::COO(const matrix::COO<T> &coo) {
 template COO<double>::COO(const matrix::COO<double> &coo);
 template COO<float>::COO(const matrix::COO<float> &coo);
 
+// operator=
+template <typename T> void COO<T>::operator=(const matrix::COO<T> &mat) {
+  Logger &logger = Logger::get_instance();
+  logger.util_in(monolish_func);
+
+  // err
+  if (get_row() != mat.get_row()) {
+    throw std::runtime_error("error A.row != C.row");
+  }
+  if (get_col() != mat.get_col()) {
+    throw std::runtime_error("error A.col != C.col");
+  }
+  if (get_nnz() != mat.get_nnz()) {
+    throw std::runtime_error("error A.nnz != C.nnz");
+  }
+  if (get_device_mem_stat() != mat.get_device_mem_stat()) {
+    throw std::runtime_error("error get_device_mem_stat() is not same");
+  }
+
+  // value copy
+  internal::vcopy(get_nnz(), val.data(), mat.val.data(), get_device_mem_stat());
+
+  logger.util_out();
+}
+
 // input and convert //////////////////////////////////////
-//
+
 template <typename T> void COO<T>::fill(T value) {
   Logger &logger = Logger::get_instance();
   logger.util_in(monolish_func);
@@ -211,7 +236,7 @@ template <typename T> void COO<T>::input_mm(const char *filename) {
 template void COO<double>::input_mm(const char *filename);
 template void COO<float>::input_mm(const char *filename);
 
-template <typename T> void COO<T>::print_all() const {
+template <typename T> void COO<T>::print_all(bool force_cpu) const {
   Logger &logger = Logger::get_instance();
   logger.util_in(monolish_func);
   std::cout << std::scientific;
@@ -228,8 +253,8 @@ template <typename T> void COO<T>::print_all() const {
   }
   logger.util_out();
 }
-template void COO<double>::print_all() const;
-template void COO<float>::print_all() const;
+template void COO<double>::print_all(bool force_cpu) const;
+template void COO<float>::print_all(bool force_cpu) const;
 
 template <typename T> void COO<T>::print_all(std::string filename) const {
   Logger &logger = Logger::get_instance();