+++ src/utils/crs_util.cpp
@@ -191,20 +191,40 @@ template <typename T> void CRS<T>::fill(T value) {
 template void CRS<double>::fill(double value);
 template void CRS<float>::fill(float value);
 
-template <typename T> void CRS<T>::print_all() {
+template <typename T> void CRS<T>::print_all(bool force_cpu) const {
   Logger &logger = Logger::get_instance();
   logger.util_in(monolish_func);
 
-  for (size_t i = 0; i < get_row(); i++) {
-    for (size_t j = (size_t)row_ptr[i]; j < (size_t)row_ptr[i + 1]; j++) {
-      std::cout << i + 1 << " " << col_ind[j] + 1 << " " << val[j] << std::endl;
+  std::cout << (MM_BANNER " " MM_MAT " " MM_FMT " " MM_TYPE_REAL
+                          " " MM_TYPE_GENERAL)
+            << std::endl;
+  std::cout << rowN << " " << colN << " " << nnz << std::endl;
+
+  if (get_device_mem_stat() == true && force_cpu == false) {
+#if MONOLISH_USE_GPU
+#pragma omp target
+    for (size_t i = 0; i < get_row(); i++) {
+      for (size_t j = (size_t)row_ptr[i]; j < (size_t)row_ptr[i + 1]; j++) {
+        printf("%d %d %f\n", i + 1, col_ind[j] + 1, val[j]);
+      }
+    }
+#else
+    throw std::runtime_error(
+        "error USE_GPU is false, but get_device_mem_stat() == true");
+#endif
+  } else {
+    for (size_t i = 0; i < get_row(); i++) {
+      for (size_t j = (size_t)row_ptr[i]; j < (size_t)row_ptr[i + 1]; j++) {
+        std::cout << i + 1 << " " << col_ind[j] + 1 << " " << val[j]
+                  << std::endl;
+      }
     }
   }
 
   logger.util_out();
 }
-template void CRS<double>::print_all();
-template void CRS<float>::print_all();
+template void CRS<double>::print_all(bool force_cpu) const;
+template void CRS<float>::print_all(bool force_cpu) const;
 
 template <typename T> bool CRS<T>::operator==(const CRS<T> &mat) const {
   Logger &logger = Logger::get_instance();