+++ test/eigen/LOBPCG/lobpcg.cpp
@@ -5,10 +5,44 @@
 #include <iostream>
 
 template <typename T, typename PRECOND>
-bool test(const char *file, const int check_ans, const T tol,
-          const int maxiter) {
-  int DIM = 12;
-  monolish::matrix::COO<T> COO = monolish::util::frank_matrix<T>(DIM);
+bool test_solve(monolish::matrix::COO<T> mat, const T exact_result,
+                const int check_ans, const T tol_ev, const T tol_res) {
+  monolish::matrix::CRS<T> A(mat);
+  T lambda;
+  monolish::vector<T> x(A.get_row());
+
+  monolish::eigen::LOBPCG<T> solver;
+
+  solver.set_tol(tol_res);
+  solver.set_lib(0);
+  solver.set_miniter(0);
+  solver.set_maxiter(1000);
+
+  // precond setting
+  PRECOND precond;
+  solver.set_create_precond(precond);
+  solver.set_apply_precond(precond);
+
+  solver.set_print_rhistory(true);
+
+  if (monolish::util::solver_check(solver.solve(A, lambda, x))) {
+    return false;
+  }
+
+  if (check_ans == 1) {
+    if (ans_check<T>("LOBPCG", lambda, exact_result, tol_ev) == false) {
+      return false;
+    }
+  }
+  return true;
+}
+
+template <typename T, typename PRECOND>
+bool test_tridiagonal_toeplitz(const int check_ans, const T tol_ev,
+                               const T tol_res) {
+  int DIM = 100;
+  monolish::matrix::COO<T> COO =
+      monolish::util::tridiagonal_toeplitz_matrix<T>(DIM, 11.0, -1.0);
   // for (std::size_t i = 0; i < COO.get_row(); ++i) {
   //   for (std::size_t j = 0; j < COO.get_col(); ++j) {
   //     if (i == j) { COO.insert(i, j, 2.0); }
@@ -30,37 +64,19 @@ bool test(const char *file, const int check_ans, const T tol,
   // T exact_result = ld[0];
 
   // Calculate exact eigenvalue from analytic solution
-  T exact_result = monolish::util::frank_matrix_eigenvalue<T>(DIM, 1);
+  T exact_result = monolish::util::tridiagonal_toeplitz_matrix_eigenvalue<T>(
+      DIM, 0, 11.0, -1.0);
 
-  monolish::matrix::CRS<T> A(COO);
+  return test_solve<T, PRECOND>(COO, exact_result, check_ans, tol_ev, tol_res);
+}
 
-  T lambda;
-  monolish::vector<T> x(A.get_row());
+template <typename T, typename PRECOND>
+bool test_laplacian_1d(const int check_ans, const T tol_ev, const T tol_res) {
+  int DIM = 10000;
+  monolish::matrix::COO<T> COO = monolish::util::laplacian_matrix_1D<T>(DIM);
+  T exact_result = monolish::util::laplacian_matrix_1D_eigenvalue<T>(DIM, 0);
 
-  monolish::eigen::LOBPCG<T> solver;
-
-  solver.set_tol(1.0e-2);
-  solver.set_lib(0);
-  solver.set_miniter(0);
-  solver.set_maxiter(maxiter);
-
-  // precond setting
-  PRECOND precond;
-  solver.set_create_precond(precond);
-  solver.set_apply_precond(precond);
-
-  solver.set_print_rhistory(true);
-
-  if (monolish::util::solver_check(solver.solve(A, lambda, x))) {
-    return false;
-  }
-
-  if (check_ans == 1) {
-    if (ans_check<T>("LOBPCG", lambda, exact_result, tol) == false) {
-      return false;
-    }
-  }
-  return true;
+  return test_solve<T, PRECOND>(COO, exact_result, check_ans, tol_ev, tol_res);
 }
 
 int main(int argc, char **argv) {
@@ -76,23 +92,30 @@ int main(int argc, char **argv) {
   // monolish::util::set_log_level(3);
   // monolish::util::set_log_filename("./monolish_test_log.txt");
 
-  if (test<double, monolish::equation::none<double>>(file, check_ans, 1.0e-0,
-                                                     150) == false) {
+  if (test_tridiagonal_toeplitz<double, monolish::equation::none<double>>(check_ans, 3.0e-2, 8.0e-2) == false) {
     return 1;
   }
-  if (test<float, monolish::equation::none<float>>(file, check_ans, 1.0e-0,
-                                                   150) == false) {
+  if (test_tridiagonal_toeplitz<float, monolish::equation::none<float>>(check_ans, 1.0e-1, 1.0e-0) == false) {
     return 1;
   }
-
-  if (test<double, monolish::equation::Jacobi<double>>(file, check_ans, 1.0e+1,
-                                                       150) == false) {
+  if (test_tridiagonal_toeplitz<double, monolish::equation::Jacobi<double>>(check_ans, 1.0e-2, 1.0e-2) == false) {
     return 1;
   }
-  if (test<float, monolish::equation::Jacobi<float>>(file, check_ans, 1.0e+1,
-                                                     90) == false) {
+  if (test_tridiagonal_toeplitz<float, monolish::equation::Jacobi<float>>(check_ans, 5.0e-2, 5.0e-2) == false) {
     return 1;
   }
 
+  if (test_laplacian_1d<double, monolish::equation::none<double>>(check_ans, 2.0e-2, 5.0e-2) == false) {
+    return 1;
+  }
+  if (test_laplacian_1d<float, monolish::equation::none<float>>(check_ans, 1.0e-1, 3.0e-1) == false) {
+    return 1;
+  }
+  if (test_laplacian_1d<double, monolish::equation::Jacobi<double>>(check_ans, 5.0e-2, 5.0e-2) == false) {
+    return 1;
+  }
+  if (test_laplacian_1d<float, monolish::equation::Jacobi<float>>(check_ans, 2.0e-1, 2.0e-1) == false) {
+    return 1;
+  }
   return 0;
 }